requirements.txt
# VideoRAG & EBR-RAG Core Requirements
# Optimized for Python 3.12 and CUDA 12.1

--index-url https://download.pytorch.org/whl/cu121

# Core Deep Learning
torch>=2.4.0
torchvision
torchaudio
accelerate==0.30.1
bitsandbytes==0.43.1
numpy==1.26.4

# Multi-modal & Vision
timm
ftfy
regex
einops
fvcore
iopath
matplotlib
types-regex
transformers==4.37.1

# Video & Audio Processing
moviepy==1.0.3
ctranslate2
faster_whisper
# Patched repositories should be installed via:
# pip install --no-deps git+https://github.com/facebookresearch/pytorchvideo.git
# pip install --no-deps git+https://github.com/facebookresearch/ImageBind.git

# Storage & Databases
neo4j
hnswlib
xxhash
nano-vectordb

# LLM & Utilities
openai
tenacity
tiktoken
ollama
python-dotenv
setuptools<70
